Splitting classes into multiple files
